project finish

+++ app/Http/Controllers/BookController.php
@@ -52,34 +52,50 @@ class BookController extends Controller
     public function store(Request $request)
     // title, description. author. publisher. date_of_issue
     {
+        $request->validate([
+            "author_id" => ['required', 'numeric'],
+        ]);
 
-        try {
-            Book::create([
-                "title" => $request->title,
-                "description" => $request->description,
-                "book_image" => $request->file('image')->store('book-image'),
-                "publisher" => $request->publisher,
-                "author_id" => $request->author_id,
-                "count_book" => $request->count_book,
-                "date_of_issue" => $request->date_of_issue,
-
-            ]);
-
-            $request->session()->flash('successTambah', 'Buku berhasil di tambah!');
-            return redirect('/');
-        } catch (Exception $ex) {
-            $request->session()->flash('gagalTambah',  $ex->getMessage());
-            return redirect('/');
+        $author = Authors::where('id', $request->author_id);
+        if (count($author->get()) == 0) {
+            $request->session()->flash('statusAuthor', 'Pengarang tidak di temukan!');
+            return redirect('/addBook');
+        } else {
+            try {
+                Book::create([
+                    "title" => $request->title,
+                    "description" => $request->description,
+                    "book_image" => $request->file('image')->store('book-image'),
+                    "publisher" => $request->publisher,
+                    "author_id" => $request->author_id,
+                    "count_book" => $request->count_book,
+                    "date_of_issue" => $request->date_of_issue,
+                ]);
+                $request->session()->flash('successTambah', 'Buku berhasil di tambah!');
+                return redirect('/');
+            } catch (Exception $ex) {
+                $request->session()->flash('gagalTambah',  $ex->getMessage());
+                return redirect('/');
+            }
         }
     }
 
     public function indexForm()
     {
-        $author = Authors::get();
-        return view('layouts.crud.addBook',  [
-            'halaman' => 'book',
-            'author' => $author
-        ]);
+        $author = Authors::latest();
+        if (count($author->get()) == 0) {
+
+            return view('layouts.crud.addBook', [
+                'halaman' => 'book',
+                'author' => "",
+            ]);
+        } else {
+            $author = Authors::get();
+            return view('layouts.crud.addBook',  [
+                'halaman' => 'book',
+                'author' => $author
+            ]);
+        }
     }
 
     /**
@@ -90,9 +106,6 @@ class BookController extends Controller
      */
     public function show($id)
     {
-        /* $data = Book::find($id);
-        $authors_id = Book::select('author_id')->where('id', $id)->get();
-        $authors = Authors::where('id', Book::pluck("author_id"))->get(); */
         $data = Book::with('authors')->find($id);
         if ($data == null) {
             return response([
@@ -106,21 +119,6 @@ class BookController extends Controller
                 'data' => $data,
             ], 200);
         }
-    }
-    public function search()
-    {
-        //dd(request('search'));
-        /* $data = Book::where('title', 'LIKE', "%$title%")->get();
-        if (count($data) > 0) {
-            return view('buku.index', [
-                'book' => $data,
-            ]);
-        } else {
-            return response([
-                'status' => 404,
-                'message' => "Tidak ada data dengan title $title",
-            ], 404);
-        } */
     }
 
     /**
@@ -145,38 +143,60 @@ class BookController extends Controller
     {
         $data = Book::where('id', $id)->first();
 
-        $author = Authors::get();
-        return view('layouts.crud.editBook',  [
-            'halaman' => 'book',
-            'author' => $author,
-            'data' => $data
-        ]);
+        $author = Authors::latest();
+        if (count($author->get()) == 0) {
+            return view('layouts.crud.editBook',  [
+                'halaman' => 'book',
+                'author' => "",
+                'data' => $data
+            ]);
+        } else {
+            $author = Authors::get();
+            return view('layouts.crud.editBook',  [
+                'halaman' => 'book',
+                'author' => $author,
+                'data' => $data
+            ]);
+        }
     }
     public function update(Request $request, $id)
     {
-
         $data = Book::find($id);
+        $validatedData = $request->validate([
+            'author_id' => ['required', 'numeric'],
+        ]);
 
-        if ($request->hasFile('image')) {
-            $request->validate([
-                'image' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
-            ]);
-            if ($data->book_image) {
-                Storage::delete($data->book_image);
+        $author = Authors::where('id', $request->author_id);
+        if (count($author->get()) == 0) {
+            $request->session()->flash('statusAuthor', 'Pengarang tidak di temukan!');
+            return redirect('/editBook/' . $id);
+        } else {
+            try {
+                if ($request->hasFile('image')) {
+                    $request->validate([
+                        'image' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
+                    ]);
+                    if ($data->book_image) {
+                        Storage::delete($data->book_image);
+                    }
+                    $path = $request->file('image')->store('book-image');
+                    $data->book_image = $path;
+                }
+                $data->title = $request->title;
+                $data->description = $request->description;
+                $data->publisher = $request->publisher;
+                if ($validatedData) {
+                    $data->author_id = $request->author_id;
+                }
+                $data->count_book = $request->count_book;
+                $data->date_of_issue = $request->date_of_issue;
+                $data->save();
+                return redirect('/')->with('successEdit', 'Buku berhasil di Edit!');
+            } catch (Exception $ex) {
+                $request->session()->flash('gagalEdit',  $ex->getMessage());
+                return redirect('/');
             }
-            $path = $request->file('image')->store('book-image');
-            $data->book_image = $path;
         }
-        $data->title = $request->title;
-        $data->description = $request->description;
-        $data->publisher = $request->publisher;
-        $data->author_id = $request->author_id;
-        $data->count_book = $request->count_book;
-        $data->date_of_issue = $request->date_of_issue;
-        $data->save();
-        return redirect('/')->with('status', 'Buku berhasil di Edit');
-        //url('/detailBook/' . $data->id)
-
     }
 
     /**
